Внес небольшие правки в код, переместил модель комнат в папку модедей отелей.
