+++ app/hotels/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import JSON, Column, Integer, String
+from sqlalchemy import JSON, Column, ForeignKey, Integer, String
 from app.database import Base
 
 
@@ -12,4 +12,19 @@ class Hotels(Base):
     location = Column(String, nullable=False)
     services = Column(JSON)
     rooms_quantity = Column(Integer, nullable=False)
+    image_id = Column(Integer)
+
+
+class Rooms(Base):
+    """Модель для таблицы комнат."""
+
+    __tablename__ = "rooms"
+
+    id = Column(Integer, primary_key=True)
+    hotel_id = Column(Integer, ForeignKey("hotels.id"))
+    name = Column(String, nullable=False)
+    description = Column(String, nullable=False)
+    price = Column(Integer, nullable=False)
+    services = Column(JSON)
+    quantity = Column(Integer, nullable=False)
     image_id = Column(Integer)
